backend/requirements.txt
fastapi
uvicorn
mercadopago
python-dotenv
